Addressed review feedback.
