+++ kernel/libc/c11/Makefile
@@ -11,6 +11,6 @@ OBJS = call_once.o cnd_broadcast.o cnd_destroy.o cnd_init.o cnd_signal.o \
     mtx_timedlock.o mtx_trylock.o mtx_unlock.o thrd_create.o thrd_current.o \
     thrd_detach.o thrd_equal.o thrd_exit.o thrd_join.o thrd_sleep.o \
     thrd_yield.o tss_create.o tss_delete.o tss_get.o tss_set.o \
-    aligned_alloc.o timespec_get.o atomics.o
+    aligned_alloc.o timespec_get.o timegm.o atomics.o
 
 include $(KOS_BASE)/Makefile.prefab
